# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cloud-api/onboarding/verify-email": {
      "post": {
        "operationId": "OnboardingController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Email verification OTP payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Email verified successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify email address using OTP",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/resend-email-otp": {
      "post": {
        "operationId": "OnboardingController_resendEmailOtp",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OTP sent successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resend email verification OTP",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/status": {
      "get": {
        "operationId": "OnboardingController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current onboarding status for the user"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current onboarding status",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/start": {
      "post": {
        "operationId": "OnboardingController_startOnboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Onboarding process started, returns current step and sends OTP if needed"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start or continue the onboarding process",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/set-password": {
      "post": {
        "operationId": "OnboardingController_setPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "New password payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password set successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid password format or validation failed"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set password for OAuth users",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/mobile-verification/initiate": {
      "post": {
        "operationId": "OnboardingController_initiateMobileVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Mobile verification initiation payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateMobileVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mobile verification initiated successfully"
          },
          "400": {
            "description": "Invalid phone number or verification method"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate mobile phone verification",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/mobile-verification/status": {
      "get": {
        "operationId": "OnboardingController_getMobileVerificationStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current mobile verification status"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          },
          "404": {
            "description": "No active mobile verification found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current mobile verification status",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/mobile-verification/resend": {
      "post": {
        "operationId": "OnboardingController_resendMobileVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Mobile verification resend payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateMobileVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mobile verification code resent successfully"
          },
          "400": {
            "description": "Invalid phone number or verification method"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resend mobile verification code",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/cloud-api/onboarding/mobile-verification/verify-otp": {
      "post": {
        "operationId": "OnboardingController_verifyMobileOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Mobile OTP verification payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMobileOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phone number verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Phone number verified successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify mobile phone number using OTP",
        "tags": [
          "Onboarding"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti Cloud API",
    "description": "Internal API for Vritti SaaS Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string",
            "description": "Six-digit one-time password sent to the user email address for verification",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "otp"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "User password that must be at least 8 characters and contain at least one lowercase letter, uppercase letter, number, and special character",
            "example": "SecureP@ss123",
            "minLength": 8
          }
        },
        "required": [
          "password"
        ]
      },
      "InitiateMobileVerificationDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Phone number in E.164 format (with + prefix). Optional for QR-based methods, required for OTP-based method.",
            "example": "+919876543210",
            "minLength": 10,
            "maxLength": 20
          },
          "phoneCountry": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2 country code for the phone number",
            "example": "IN",
            "minLength": 2,
            "maxLength": 5
          },
          "method": {
            "type": "string",
            "description": "Method to use for mobile verification. Defaults to WHATSAPP_QR.",
            "example": "WHATSAPP_QR",
            "enum": [
              "WHATSAPP_QR",
              "SMS_QR",
              "MANUAL_OTP"
            ],
            "default": "WHATSAPP_QR"
          }
        }
      },
      "VerifyMobileOtpDto": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string",
            "description": "Six-digit one-time password received via SMS for mobile verification",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "otp"
        ]
      },
...
    }
  }
}